Fixed cql insert, now inserts stuff

+++ src/main/java/fi/hsl/transitlog/hfp/HfpMessage.java
@@ -27,7 +27,7 @@ public class HfpMessage {
 
         public String dir;
 
-        public Integer oper;
+        public Short oper;
 
         public Integer veh;
